% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rasterizer.R
\name{add_rasterizer}
\alias{add_rasterizer}
\title{Add "rasterizer" trace to a plotly visualization}
\usage{
add_rasterizer(p, x = NULL, y = NULL, z = NULL, ..., data = NULL,
  inherit = TRUE, on = NULL, size = NULL, scaling = c("log",
  "to01", "origin"))
}
\arguments{
\item{x}{The x variable, numerical vectors or expression and will be passed on `aes()`.}

\item{y}{The y variable, numerical vectors or expression and will be passed on `aes()`.}

\item{z}{A numeric matrix, if provided, `add_heatmap` will be called.}

\item{...}{Arguments (i.e., attributes) passed along to the trace type or `rasterizer`.}

\item{data}{A data frame (optional) or \link{crosstalk::SharedData} object.}

\item{inherit}{Inherit attributes from \link{plot_ly()}}

\item{on}{Reduction "on" which variable, it is a numerical vectors or expression and will be passed on `aes()`.}

\item{size}{The size of pixel for each observation, a numerical vectors or expression and will be passed on `aes()`.}

\item{scaling}{Scale layout matrix.}
}
\description{
Add trace to a plotly visualization
}
\examples{
\dontrun{
   library(rasterizer)
   if(requireNamespace("plotly")) {
     x <- rnorm(1e7)
     y <- rnorm(1e7)
     min_x <- min(x)
     min_y <- min(y)
     max_x <- max(x)
     max_y <- max(y)
     data.table::data.table(x = x, y = y) \%>\% 
       plot_ly() \%>\% 
       add_rasterizer(x = ~x, y = ~y,
                      x_range = c(min_x, max_x), y_range = c(min_y, max_y))
       
   }
}
}
