% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_points.R
\name{aggregation_points}
\alias{aggregation_points}
\title{aggregation_points}
\usage{
aggregation_points(rastObj, data = NULL, mapping = aes(), ...,
  xlim = NULL, ylim = NULL, max_size = NULL, reduction_func = NULL,
  layout = NULL, glyph = NULL, group_by_data_table = NULL)
}
\arguments{
\item{rastObj}{A `rasterObj`.}

\item{reduction_func}{A reduction operator function is used when aggregating datapoints
into a given pixel. The supported reduction operators are `sum`, `any`, `mean`, `first` and `last` so far.
More is coming}

\item{layout}{The way to layout multiple images. It is useful when data is catergorical("colour" is set in `aes()`).
The default setting is "weighted", which means the final raster is the weighted combination of each category
aggregation matrix. Also, we can "cover" each raster by the order of unique categories.}
}
\description{
points layer for `rasterizer`
}
