% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorMaps.R
\docType{data}
\name{color_map}
\alias{color_map}
\alias{fire_map}
\alias{viridis_map}
\alias{hourColors_map}
\title{Supplemental color maps for rasterly}
\format{An object of class \code{character} of length 256.}
\usage{
fire_map

viridis_map

hourColors_map
}
\description{
Hex codes for the color map.
Used in setting argument \code{color} in \code{rasterly} or \code{rasterly} layers.
}
\keyword{datasets}
