% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourMap.R
\docType{data}
\name{fire}
\alias{fire}
\alias{viridis}
\alias{hourColours}
\title{Color map}
\format{An object of class \code{character} of length 256.}
\usage{
fire

viridis

hourColours
}
\description{
A vector of colors. 
Used in setting argument `color_map` or `color_key` in \code{rasterly} or \code{rasterly} layers.
They may be deprecated in the future.
}
\keyword{datasets}
