% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggRasterly.R
\name{ggRasterly}
\alias{ggRasterly}
\title{ggRasterly}
\usage{
ggRasterly(data = NULL, mapping = aes(), ..., plot_width = 600,
  plot_height = 600, x_range = NULL, y_range = NULL,
  background = "white", color = NULL, show_raster = TRUE,
  drop_data = FALSE, variable_check = FALSE, alpha = 0.5,
  shape = 19, size = 0.5)
}
\arguments{
\item{alpha}{The transparency of points, from 0 to 1.}

\item{shape}{The shape of points, see \link{pch}.}

\item{size}{The size of points.}
}
\value{
a `ggplot` object
}
\description{
Display large data set in \code{ggplot}.
}
\examples{
\dontrun{
if(requireNamespace("ggplot2") && requireNamespace("data.table") && 
  requireNamespace("lubridate")) {
 # Load data
 url1 <- "https://raw.githubusercontent.com/plotly/datasets/master/uber-rides-data1.csv"
 ridesRaw_1 <-  url1 \%>\%
   data.table::fread(stringsAsFactors = FALSE)
 url2 <- "https://raw.githubusercontent.com/plotly/datasets/master/uber-rides-data2.csv"
 ridesRaw_2 <-  url2 \%>\%
   data.table::fread(stringsAsFactors = FALSE)
 url3 <- "https://raw.githubusercontent.com/plotly/datasets/master/uber-rides-data3.csv"
 ridesRaw_3 <-  url3 \%>\%
   data.table::fread(stringsAsFactors = FALSE) 
   
 ridesDf <- list(ridesRaw_1, ridesRaw_2, ridesRaw_3) \%>\%
   data.table::rbindlist()
   
 time <- lubridate::ymd_hms(ridesDf$`Date/Time`)
 ridesDf <-  ridesDf[, 'Date/Time':=NULL][, list(Lat,
                                                 Lon,
                                                 hour = lubridate::hour(time),
                                                 month = lubridate::month(time),
                                                 day = lubridate::day(time))]

 #### quick start
 ggRasterly(data = ridesDf, 
            mapping = aes(x = Lat, y = Lon, color = hour),
            color = hourColors
 ) + 
 labs(title = "New York Uber",
      subtitle = "Apr to Sept, 2014",
      caption = "Data from https://raw.githubusercontent.com/plotly/datasets/master")
 }
}
}
\seealso{
\link{plotRasterly}
}
