% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorMaps.R
\docType{data}
\name{hourColors_map}
\alias{hourColors_map}
\title{Supplemental color maps for rasterly}
\format{An object of class \code{character} of length 24.}
\usage{
hourColors_map
}
\description{
Hex codes for the "hourColors" color map.
Used in setting argument `color_map` or `color_key` in \code{rasterly} or \code{rasterly} layers.
This may be deprecated in the future.
}
\keyword{datasets}
