% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_rasterly.R
\name{is.rasterlyBuild}
\alias{is.rasterlyBuild}
\title{Is \code{rasterlyBuild}}
\usage{
is.rasterlyBuild(x)
}
\arguments{
\item{x}{a \code{rasterly} object}
}
\description{
Reports whether x is a \code{rasterlyBuild} object. In other word, it helps to define
whether this object has been passed through `rasterly_build`
}
