% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterizerGrob.R
\name{plot.rasterizer}
\alias{plot.rasterizer}
\title{Plot "rasterizer" object.}
\usage{
\method{plot}{rasterizer}(x, y = NULL, ...)
}
\arguments{
\item{x}{a "rasterizer" object to be plotted}

\item{y}{NULL, will be ignored.}

\item{...}{parameters passed to \code{grid.rasterizer} and \code{rasterizerGrob}}
}
\value{
invisible()
}
\description{
This is a wrapper for \code{grid.rasterizer()} to simplify the plotting.
}
\examples{
mtcars \%>\% 
  rasterizer(plot_height = 20, 
             plot_width = 20, 
             mapping = aes(x = hp, y = mpg, colour = cyl)) \%>\% 
  rasterize_points() \%>\% 
  execute() -> p
  plot(x = p)
}
\seealso{
\code{\link{grid.rasterizer}}, \code{\link{rasterizerGrob}}
}
