% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.rasterizer.R
\name{plotly.rasterizer}
\alias{plotly.rasterizer}
\title{`Rasterizer` to `plotly`}
\usage{
plotly.rasterizer(rastObj, as_heatmap = FALSE, scaling = c("log",
  "to01", "origin"), ...)
}
\arguments{
\item{rastObj}{A rasterizer object}

\item{as_heatmap}{Draw `plotly` by adding heatmap layer. See \code{add_rasterizer}}

\item{scaling}{Scale layout matrix.}

\item{...}{Arguments to the layout object. For documentation, 
see https://plot.ly/r/reference/#Layout_and_layout_style_objects}
}
\description{
Display raster image via `plotly`
}
\examples{
\dontrun{
   library(rasterizer)
   if(requireNamespace("plotly")) {
     x <- rnorm(1e7)
     y <- rnorm(1e7)
     category <- sample(1:5, 1e7, replace = TRUE)
     data.frame(x = x, y = y, category = category) \%>\%
       canvas(mapping = aes(x = x, y = y)) \%>\%
       aggregation_points(layout = "weighted") \%>\%
       rasterizer() \%>\% 
       plotly.rasterizer()
   }
}
}
