% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterizer_build.R
\name{rasterizer_build}
\alias{rasterizer_build}
\title{rasterizer_build}
\usage{
rasterizer_build(rastObj)
}
\arguments{
\item{rastObj}{A rasterizer object. It should be a list of environments composed of a `rasterizer()` and 
several `rasterize_...` layers}
}
\description{
Execute a "rasterizer" object and return the info to build image
}
\note{
A rasterizer object will never be implemented until we call `rasterizer_build()`
}
\examples{
r <- data.frame(x = rnorm(1e5), y = rnorm(1e5)) \%>\% 
       rasterizer(mapping = aes(x = x, y = y)) \%>\%
       rasterize_points(colour_map = fire)
str(r)
p <- rasterizer_build(r)
str(p)
}
\seealso{
\link{rasterizer}, \link{rasterize_points}, \link{[.rasterizer}, \link{[<-.rasterizer}
}
