% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterizerGrob.R
\name{rasterizerGrob}
\alias{rasterizerGrob}
\title{Render a "rasterizer" object}
\usage{
grid.rasterizer(rastObj, ..., interpolate = FALSE,
  name = NULL, gp = grid::gpar(), vp = NULL)
    
rasterizerGrob(rastObj, ..., interpolate = FALSE, 
  name = NULL, gp = grid::gpar(), vp = NULL)
}
\arguments{
\item{rastObj}{A "rasterizer" object}

\item{...}{To specify layout aesthetics. See more in details}

\item{interpolate}{A logical value indicating whether to linearly interpolate the image 
(the alternative is to use nearest-neighbour interpolation, which gives a more blocky result).}

\item{name}{A character identifier.}

\item{gp}{An object of class gpar, typically the output from a call to the function gpar. 
This is basically a list of graphical parameter settings.}

\item{vp}{A Grid viewport object (or NULL).}
}
\description{
Render a "rasterizer" object
}
\details{
Set ... in `grid.rasterizer()` and `rasterizerGrob()`
\itemize{
 \item{}{title: The title of output graph}
 \item{}{title_fontsize: title fontsize}
 \item{}{title_fontfamily: title fontfamily}
 \item{}{title_fontface: title fontface}
 \item{}{title_location: title location}
 \item{}{title_hvjust: title hvjust, a length 2 vector represents "hjust" and "vjust" correspondingly}
 \item{}{title_col: title colour}
 \item{}{xlabel: x label, default setting is the "x" name in mapping system}
 \item{}{ylabel: y label, default setting is the "y" name in mapping system}
 \item{}{label_fontsize: label fontsize}
 \item{}{label_fontfamily: label fontfamily}
 \item{}{label_fontface: label fontface}
 \item{}{label_location: label location}
 \item{}{label_hvjust: label hvjust, a length 2 vector represents "hjust" and "vjust" correspondingly}
 \item{}{label_col: label colour}
 \item{}{axis_col: axis colour}
 \item{}{axis_fontsize: axis fontsize}
 \item{}{image_width: image width}
 \item{}{image_height: image height}
}
}
\examples{
iris \%>\% 
  rasterizer(plot_height = 20, 
             plot_width = 20, 
             mapping = aes(x = Sepal.Width, y = Petal.Length, colour = Species)) \%>\% 
  rasterize_points() \%>\% 
  rasterizer_build() \%>\% 
  grid.rasterizer()
}
\seealso{
\code{\link{plot.rasterizer}}, \code{\link{plotly.rasterizer}}
}
