\docType{package}
\name{rasterly-package}
\alias{rasterly}
\alias{rasterly-package}
\title{Easily and rapidly generate raster image data for large datasets using Plotly.js}
\description{
rasterly is an R package to generate raster data for very large datasets; combined with Plotly.js and the \link{plotly} package, it enables analysts to generate interactive figures which are responsive enough to embed into web applications.
}
\seealso{
\itemize{
  \item \url{https://github.com/plotly/rasterly}
  \item Report bugs at \url{https://github.com/plotly/rasterly/issues}
}

}
\author{
\strong{Maintainer}: Zehao Xu \email{z267xu@uwaterloo.ca}

Authors:
\itemize{
  \item Zehao Xu \email{z267xu@uwaterloo.ca}
  \item Ryan Patrick Kyle \email{ryan@plot.ly}
}

Other contributors:
\itemize{
  \item  Plotly Technologies [copyright holder]
}

}