% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterly_build.R
\name{rasterly_build}
\alias{rasterly_build}
\title{rasterly_build}
\usage{
rasterly_build(rastObj)
}
\arguments{
\item{rastObj}{A rasterly object. It should be a list of environments composed of a `rasterly()` and
several `rasterly_...` layers}
}
\description{
Execute a "rasterly" object and return the info to build image.
}
\note{
A rasterly object will never be implemented until we call `rasterly_build()`
}
\examples{
r <- data.frame(x = rnorm(1e5), y = rnorm(1e5)) \%>\%
       rasterly(mapping = aes(x = x, y = y)) \%>\%
       rasterize_points(color_map = fire)
str(r)
p <- rasterly_build(r)
str(p)
}
\seealso{
\link{rasterly}, \link{rasterize_points}, \link{[.rasterly}, \link{[<-.rasterly}
}
