% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterly_guides.R
\name{rasterly_guides}
\alias{rasterly_guides}
\title{rasterly_guides}
\usage{
rasterly_guides(
  rastObj,
  x_pretty = NULL,
  y_pretty = NULL,
  panel_background = "grey92",
  panel_line = "white"
)
}
\arguments{
\item{rastObj}{A "rasterly" object.}

\item{x_pretty}{The pretty on x. Compute a sequence of about n+1 equally spaced 'round' values 
which cover the range of the values in x. If it is not provided, \code{x_pretty} will 
be generated by the x range}

\item{y_pretty}{The pretty on y.}

\item{panel_background}{Panel background.}

\item{panel_line}{Panel line color}
}
\description{
Guides layer for "rasterly".
}
\details{
When an image has a 'complicated' background, the drawing time increases significantly.
So it is not recommended. A suggestion to draw grid guides is to transform image data to a data frame 
via \link{image2data}, then use \code{ggplot} or \code{plotly} to display.
}
\seealso{
\link{ggRasterly}
}
