% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourMap.R
\docType{data}
\name{fire}
\alias{fire}
\alias{viridis}
\alias{hourColors}
\title{fire}
\format{An object of class \code{character} of length 256.}
\usage{
fire

viridis

hourColors
}
\description{
Hex codes for the "fire" color map.
Used in setting argument `color_map` or `color_key` in \code{rasterly} or \code{rasterly} layers.
These may be deprecated in the future.

Hex codes for the "viridis" color map.

Hex codes for a palette with 24 colors, one for each hour of the day.
}
\keyword{datasets}
