% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_points.R
\name{rasterize_points}
\alias{rasterize_points}
\title{rasterize_points}
\usage{
rasterize_points(rastObj, data = NULL, mapping = aes(), ...,
  xlim = NULL, ylim = NULL, max_size = NULL, reduction_func = NULL,
  layout = NULL, glyph = NULL, group_by_data_table = NULL,
  inherit.aes = TRUE)
}
\arguments{
\item{rastObj}{A "rasterly" object.}

\item{data}{A `data.frame` or `function` with an argument `x`, specifying the dataset to use for plotting. If `data` 
is `NULL`, the `data` argument provided to `rasterly()` may be passed through.}

\item{mapping}{Default list of aesthetic mappings to use for plot. If provided and `inherit.aes = TRUE`, it will be
stacked on top of the mappings passed to `rasterly()`.}

\item{...}{Pass-through arguments provided by `rasterly()`.}

\item{xlim}{Vector of type numeric. X limits in this layer.}

\item{ylim}{Vector of type numeric. Y limits in this layer.}

\item{max_size}{Numeric. When size changes, the upper bound of the number of pixels over which to spread a single observation.}

\item{reduction_func}{Function. A reduction function is used to aggregate data points into their pixel representations. Currently
supported reduction operators are `sum`, `any`, `mean`, `m2`, `first`, `last`, `min` and `max`. Default is `sum`. See details.}

\item{layout}{Character. The method used to generate layouts for multiple images. The default is `weighted`. Useful for categorical
data (i.e. "color" is provided via `aes()`). `weighted` specifies that the final raster should be a weighted combination of each
(categorical) aggregation matrix.}

\item{glyph}{Character. Currently, only "circle" and "square" are supported; as the `size` of the pixels increases, how should they
spread out -- should the pattern be circular or square? Other glyphs may be added in the future.}

\item{group_by_data_table}{Logical. Default is `TRUE`; when "color" is provided via `aes()`, the "group by" operation may be
perfromed within `data.table` or natively within `rasterly`. Generally, `group_by_data_table = TRUE` is faster, but for very
large datasets grouping within `rasterly` may offer better performance.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than combining with them.}
}
\description{
Points layer for "rasterly". Deprecated now, please use \code{rasterly_points} instead.
}
\seealso{
\link{rasterly_points}
}
