% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterly-package.R
\docType{package}
\name{rasterly-package}
\alias{rasterly-package}
\alias{_PACKAGE}
\title{rasterly: Easily and Rapidly Generate Raster Image Data with Support for \code{Plotly.js}}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

This package aims to easily and rapidly generate raster data in R, even for very large datasets, with an aesthetics-based mapping syntax that should be familiar to users of the \code{ggplot2} package. While \code{rasterly} does not attempt to reproduce the full functionality of the \code{Datashader} graphics pipeline system for Python, the \code{rasterly} API has several core elements in common with that software package.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/plotly/rasterly/issues}
}

}
\author{
\strong{Maintainer}: Zehao Xu \email{z267xu@uwaterloo.ca}

Other contributors:
\itemize{
  \item Ryan Patrick Kyle \email{ryan@plot.ly} (0000-0001-5829-9867) [contributor]
  \item  Plotly Technologies [copyright holder]
}

}
\keyword{internal}
