% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterly.R
\name{rasterly}
\alias{rasterly}
\title{Rasterly}
\usage{
rasterly(data = NULL, mapping = aes(), ..., plot_width = 600,
  plot_height = 600, x_range = NULL, y_range = NULL,
  background = "white", color_map = c("lightblue", "darkblue"),
  color_key = NULL, show_raster = TRUE, drop_data = FALSE,
  variable_check = FALSE)
}
\arguments{
\item{data}{Dataset to use for plot. If not provided, data must be supplied in each layer of the plot.
Since "rasterly" is used for large dataset, "\link[data.table]{data.table}" is recommended.}

\item{mapping}{Default list of aesthetic mappings to use for plot. The same with `ggplot2` \link[ggplot2]{aes}.
See details.}

\item{...}{Other arguments which will be passed through to layers.}

\item{plot_width}{Integer. The width of the image to plot; must be a positive integer. A higher value indicates a higher resolution.}

\item{plot_height}{Integer. The height of the image to plot; must be a positive integer. A higher value indicates a higher resolution.}

\item{x_range}{Vector of type numeric. The range of `x`; it can be used to clip the image. For larger datasets, providing `x_range`
may result in improved performance.}

\item{y_range}{Vector of type numeric. The range of `y`; it can be used to clip the image. For larger datasets, providing `y_range`
may result in improved performance.}

\item{background}{Character. The background color of the image to plot.}

\item{color_map}{Vector of type character. Color(s) used to draw each pixel. The `color_map` is extended by linear interpolation
independently for RGB. The darkness of the mapped color depends upon the values of the aggregation matrix.}

\item{color_key}{Vector of type character. The `color_key` is used for categorical variables; it is passed when the `color` aesthetic
is provided.}

\item{show_raster}{Logical. Should the raster be displayed?}

\item{drop_data}{Logical. When working with large datasets, drops the original data once processed according to the provided
`aes()` parameters, using the `remove()` function. See details for additional information.}

\item{variable_check}{Logical. If `TRUE`, drops unused columns to save memory; may result in reduced performance.}
}
\value{
An environment wrapped by a list
}
\description{
Set initials for rasterly.
}
\note{
Calling `rasterly()` without providing `rasterly_...()` layers has no effect.
More info can be found in \href{https://github.com/plotly/rasterly/blob/master/README.md}{README.md}
}
\seealso{
\link{rasterize_points}, \link{rasterly_build}, \link{[.rasterly}, \link{[<-.rasterly}
}
