% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{static}
\alias{static}
\alias{rasterlyGrob}
\alias{grid.rasterly}
\alias{plot.rasterly}
\alias{print.rasterly}
\title{Static rasterly graphics}
\usage{
rasterlyGrob(
  rasterlyObj,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  sub = NULL,
  interpolate = FALSE,
  axes = TRUE,
  legend = TRUE,
  legendlabel = NULL,
  legendlayer = 1,
  legendmain = NULL,
  axes_gpar = grid::gpar(col = "black", cex = 1),
  label_gpar = grid::gpar(col = "black", cex = 1),
  main_gpar = grid::gpar(col = "black", cex = 1.5),
  legend_gpar = grid::gpar(col = "black", cex = 1.5),
  name = NULL,
  gp = NULL,
  vp = NULL
)

grid.rasterly(
  rasterlyObj,
  interpolate = FALSE,
  axes = TRUE,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  sub = NULL,
  legend = TRUE,
  legendlabel = NULL,
  legendlayer = 1,
  legendmain = NULL,
  axes_gpar = grid::gpar(col = "black", cex = 1),
  label_gpar = grid::gpar(col = "black", cex = 1),
  main_gpar = grid::gpar(col = "black", cex = 1.5),
  legend_gpar = grid::gpar(col = "black", cex = 1.5),
  name = NULL,
  gp = NULL,
  vp = NULL,
  ...
)

\method{plot}{rasterly}(
  x,
  y = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  legendmain = NULL,
  sub = NULL,
  interpolate = FALSE,
  axes = TRUE,
  legend = TRUE,
  legendlabel = NULL,
  legendlayer = 1,
  new.page = TRUE,
  ...
)

\method{print}{rasterly}(x, ...)
}
\arguments{
\item{xlim}{The x limits (x1, x2) of the plot.}

\item{ylim}{The y limits (x1, x2) of the plot.}

\item{xlab}{A y label for the plot.}

\item{ylab}{A x label for the plot.}

\item{main}{A title for the plot.}

\item{sub}{A sub title for the plot.}

\item{interpolate}{A logical value indicating whether to linearly interpolate the image
(the alternative is to use nearest-neighbour interpolation, which gives a more blocky result).}

\item{axes}{Show axes or not.}

\item{legend}{Show legend or not.}

\item{legendlabel}{The label of legend.}

\item{legendlayer}{The legend represents which layer in \code{rasterly} object. Default is 1.}

\item{legendmain}{The main title of legend}

\item{axes_gpar}{\code{\link{gpar}} of axes. It is used to modify the axes' color, size and other aesthetics attributes.}

\item{label_gpar}{\code{\link{gpar}} of label. It is used to modify the labels' color, size and other aesthetics attributes.}

\item{main_gpar}{\code{\link{gpar}} of main. It is used to modify the main title's color, size and other aesthetics attributes.}

\item{legend_gpar}{\code{\link{gpar}} of legend It is used to modify the main title's color, size and other aesthetics attributes.}

\item{name}{a character identifier for the grob. Used to find the grob on the display list and/or as a child of another grob.}

\item{gp}{A gpar object, typically the output from a call to the function \code{gpar}. This is basically a list of graphical parameter settings.}

\item{vp}{a \code{\link{viewport}} object (or \code{NULL}).}

\item{...}{Other arguments to modify the display.}

\item{x}{A \code{rasterly} object}

\item{y}{NULL, will be ignored.}

\item{new.page}{display on a new page or not.}

\item{raterlyObj}{A \code{rasterly} object.}
}
\description{
Render a \code{rasterly} object. 
This object can involve axes, legend, etc to better convert graphics to data.
}
\details{
We provide three functions to produce static graphics, which is based on the API of `grid`, `plot` and `print`.
\itemize{
 \item{\code{grid}: The \code{rasterlyGrob} and \code{grid.rasterly} are the most flexible data structure. 
 It is nothing but a **\code{grob}**. Users can modify the existing display by the functions provided by \code{grid}}.
 \item{\code{plot.rasterly}: It is a S3 method. The usage is very similar to the classic `plot` function. 
 Users can set the x limits, y limits, x label, y label and etc.}
 \item{\code{print.rasterly}: It is a S3 method, however, it only provides the most basic image raster.}
}
}
